Added useGroups and useGroupDetails hooks and making code more readable
